+++ src/test/java/inshining/virtualstorage/storage/service/FolderStorageServiceTest.java
@@ -1,6 +1,6 @@
 package inshining.virtualstorage.storage.service;
 
-import inshining.virtualstorage.service.storage.FolderLocalStorageService;
+import inshining.virtualstorage.service.storage.LocalStorageService;
 import inshining.virtualstorage.util.FileDeletor;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
@@ -17,7 +17,7 @@ import static org.junit.jupiter.api.Assertions.*;
 @EnabledOnOs({OS.MAC, OS.LINUX})
 public class FolderStorageServiceTest {
     private static final String storageLocation = "upload/";
-    private static final FolderLocalStorageService folderLocalStorageService = new FolderLocalStorageService();
+    private static final LocalStorageService LocalStorageService = new LocalStorageService();
 
     private static final String USERNAME = "testUser";
 
@@ -48,7 +48,7 @@ public class FolderStorageServiceTest {
     @DisplayName("폴더 1개 만들기")
     @Test
     void createFolderTest(){
-        assertTrue(folderLocalStorageService.createFolder(USERNAME, "testFolder"));
+        assertTrue(LocalStorageService.createFolder(USERNAME, "testFolder"));
 
         Path path = Paths.get(storageLocation, USERNAME, "testFolder");
         assertTrue(Files.exists(path));
@@ -57,7 +57,7 @@ public class FolderStorageServiceTest {
     @DisplayName("존재하지 않은 하위 폴더 모두 만들기")
     @Test
     void createSubFoldersTest(){
-        assertTrue(folderLocalStorageService.createFolder(USERNAME, "testFolder1/testFolder2/testFolder3"));
+        assertTrue(LocalStorageService.createFolder(USERNAME, "testFolder1/testFolder2/testFolder3"));
 
         Path path = Paths.get(storageLocation, USERNAME, "testFolder1/testFolder2/testFolder3");
         assertTrue(Files.exists(path));
@@ -67,32 +67,32 @@ public class FolderStorageServiceTest {
     @DisplayName("이미 존재하는 폴더 만들기")
     @Test
     void createExistingFolderTest(){
-        assertTrue(folderLocalStorageService.createFolder(USERNAME, "testFolder"));
+        assertTrue(LocalStorageService.createFolder(USERNAME, "testFolder"));
 
         Path path = Paths.get(storageLocation, USERNAME, "testFolder");
         assertTrue(Files.exists(path));
 
-        assertFalse(folderLocalStorageService.createFolder(USERNAME,"testFolder"));
+        assertFalse(LocalStorageService.createFolder(USERNAME,"testFolder"));
 
         FileDeletor.delete(path, 2);
 
-        assertTrue(folderLocalStorageService.createFolder(USERNAME,"testFolder/testFolder2"));
+        assertTrue(LocalStorageService.createFolder(USERNAME,"testFolder/testFolder2"));
 
         Path path2 = Paths.get(storageLocation, USERNAME,"testFolder/testFolder2");
         assertTrue(Files.exists(path2));
 
-        assertFalse(folderLocalStorageService.createFolder(USERNAME,"testFolder"));
+        assertFalse(LocalStorageService.createFolder(USERNAME,"testFolder"));
     }
 
     @DisplayName("폴더 이름 변경")
     @Test
     void changeFolderNameTest(){
-        assertTrue(folderLocalStorageService.createFolder(USERNAME, FOLDER_NAME ));
+        assertTrue(LocalStorageService.createFolder(USERNAME, FOLDER_NAME ));
 
         Path path = Paths.get(storageLocation, USERNAME, FOLDER_NAME);
         assertTrue(Files.exists(path));
 
-        assertTrue(folderLocalStorageService.renameFolderName(USERNAME, FOLDER_NAME, CHANGED_FOLDER_NAME));
+        assertTrue(LocalStorageService.renameFolderName(USERNAME, FOLDER_NAME, CHANGED_FOLDER_NAME));
 
         Path changedPath = Paths.get(storageLocation, USERNAME, CHANGED_FOLDER_NAME);
         assertTrue(Files.exists(changedPath));
@@ -102,47 +102,47 @@ public class FolderStorageServiceTest {
     @Test
     void noExistFolderNameTest(){
         String noExistFolderName = "noExistFolder";
-        assertTrue(folderLocalStorageService.createFolder(USERNAME, FOLDER_NAME ));
+        assertTrue(LocalStorageService.createFolder(USERNAME, FOLDER_NAME ));
 
         Path path = Paths.get(storageLocation, USERNAME, FOLDER_NAME);
         assertTrue(Files.exists(path));
 
-        assertFalse(folderLocalStorageService.renameFolderName(USERNAME, noExistFolderName, CHANGED_FOLDER_NAME));
+        assertFalse(LocalStorageService.renameFolderName(USERNAME, noExistFolderName, CHANGED_FOLDER_NAME));
     }
 
     @DisplayName("실패: 존재하지 않는 유저의 폴더 이름 변경")
     @Test
     void noExistUsernameTest(){
         String noExistUsername = "noExistUser";
-        assertTrue(folderLocalStorageService.createFolder(USERNAME, FOLDER_NAME ));
+        assertTrue(LocalStorageService.createFolder(USERNAME, FOLDER_NAME ));
 
         Path path = Paths.get(storageLocation, USERNAME, FOLDER_NAME);
         assertTrue(Files.exists(path));
 
-        assertFalse(folderLocalStorageService.renameFolderName(noExistUsername, FOLDER_NAME, CHANGED_FOLDER_NAME));
+        assertFalse(LocalStorageService.renameFolderName(noExistUsername, FOLDER_NAME, CHANGED_FOLDER_NAME));
     }
 
     @DisplayName("성공: 단일 폴더 삭제")
     @Test
     void deleteFolderTest(){
-        assertTrue(folderLocalStorageService.createFolder(USERNAME, FOLDER_NAME));
+        assertTrue(LocalStorageService.createFolder(USERNAME, FOLDER_NAME));
 
         Path path = Paths.get(storageLocation, USERNAME, FOLDER_NAME);
         assertTrue(Files.exists(path));
 
-        assertTrue(folderLocalStorageService.deleteFolder(USERNAME, FOLDER_NAME));
+        assertTrue(LocalStorageService.deleteFolder(USERNAME, FOLDER_NAME));
         assertFalse(Files.exists(path));
     }
 
     @DisplayName("성공: 하위 폴더 모두 삭제")
     @Test
     void deleteSubFoldersTest(){
-        assertTrue(folderLocalStorageService.createFolder(USERNAME, "testFolder1/testFolder2/testFolder3"));
+        assertTrue(LocalStorageService.createFolder(USERNAME, "testFolder1/testFolder2/testFolder3"));
 
         Path path = Paths.get(storageLocation, USERNAME, "testFolder1/testFolder2/testFolder3");
         assertTrue(Files.exists(path));
 
-        assertTrue(folderLocalStorageService.deleteFolder(USERNAME, "testFolder1"));
+        assertTrue(LocalStorageService.deleteFolder(USERNAME, "testFolder1"));
         assertFalse(Files.exists(path));
     }
 
@@ -150,7 +150,7 @@ public class FolderStorageServiceTest {
     @Test
     void noExistFolderDeleteTest(){
         String noExistFolderName = "noExistFolder";
-        assertFalse(folderLocalStorageService.deleteFolder(USERNAME, noExistFolderName));
+        assertFalse(LocalStorageService.deleteFolder(USERNAME, noExistFolderName));
     }
 
     @DisplayName("성공: 폴더 이동 하기 (파일 이동)")
@@ -161,7 +161,7 @@ public class FolderStorageServiceTest {
         String destinationPath = "testFolder/";
         String filename = "testFile.txt";
 
-        assertTrue(folderLocalStorageService.createFolder(USERNAME, sourcePath));
+        assertTrue(LocalStorageService.createFolder(USERNAME, sourcePath));
 
         Path source = Paths.get(storageLocation, USERNAME, sourcePath);
         assertTrue(Files.exists(source));
@@ -176,10 +176,10 @@ public class FolderStorageServiceTest {
 
         Path destination = Paths.get(storageLocation, USERNAME, destinationPath);
 
-        assertTrue(folderLocalStorageService.createFolder(USERNAME, destinationPath));
+        assertTrue(LocalStorageService.createFolder(USERNAME, destinationPath));
 
         //when
-        assertTrue(folderLocalStorageService.move(USERNAME, file, destination));
+        assertTrue(LocalStorageService.move(USERNAME, file, destination));
 
         assertTrue(Files.exists(destination));
     }
@@ -192,8 +192,8 @@ public class FolderStorageServiceTest {
         String destinationPath = "testFolder4/testFolder5/testFolder6/";
         String filename = "testFile.txt";
 
-        assertTrue(folderLocalStorageService.createFolder(USERNAME, sourcePath));
-        assertTrue(folderLocalStorageService.createFolder(USERNAME, destinationPath));
+        assertTrue(LocalStorageService.createFolder(USERNAME, sourcePath));
+        assertTrue(LocalStorageService.createFolder(USERNAME, destinationPath));
 
         Path source = Paths.get(storageLocation, USERNAME, sourcePath);
         Path dest = Paths.get(storageLocation, USERNAME, destinationPath);
@@ -209,7 +209,7 @@ public class FolderStorageServiceTest {
         }
 
         //when
-        assertTrue(folderLocalStorageService.move(USERNAME, file, dest));
+        assertTrue(LocalStorageService.move(USERNAME, file, dest));
 
         // then
         assertFalse(Files.exists(file));
@@ -228,7 +228,7 @@ public class FolderStorageServiceTest {
         String destinationPath = "testFolder4/testFolder5/testFolder6/";
         String filename = "testFile.txt";
 
-        assertTrue(folderLocalStorageService.createFolder(USERNAME, sourcePath));
+        assertTrue(LocalStorageService.createFolder(USERNAME, sourcePath));
 
         Path source = Paths.get(storageLocation, USERNAME, sourcePath);
         assertTrue(Files.exists(source));
@@ -244,7 +244,7 @@ public class FolderStorageServiceTest {
         Path dest = Paths.get(storageLocation, USERNAME, destinationPath);
 
         // when
-        assertTrue(folderLocalStorageService.move(USERNAME, file, dest));
+        assertTrue(LocalStorageService.move(USERNAME, file, dest));
 
         // then
         assertFalse(Files.exists(file));
@@ -261,14 +261,14 @@ public class FolderStorageServiceTest {
         String destinationPath = "testFolder/";
         String filename = "testFile.txt";
 
-        assertTrue(folderLocalStorageService.createFolder(USERNAME, sourcePath));
+        assertTrue(LocalStorageService.createFolder(USERNAME, sourcePath));
 
         Path destination = Paths.get(storageLocation, USERNAME, destinationPath);
-        assertTrue(folderLocalStorageService.createFolder(USERNAME, destinationPath));
+        assertTrue(LocalStorageService.createFolder(USERNAME, destinationPath));
 
         Path file = Paths.get(storageLocation, USERNAME, filename);
 
-        assertFalse(folderLocalStorageService.move(USERNAME, file, destination));
+        assertFalse(LocalStorageService.move(USERNAME, file, destination));
     }
 
     @DisplayName("성공: 단일 폴더 이동 하기 (폴더 이동)")
@@ -278,8 +278,8 @@ public class FolderStorageServiceTest {
         String sourcePath = "testFolder1/";
         String destinationPath = "testFolder2/";
 
-        assertTrue(folderLocalStorageService.createFolder(USERNAME, sourcePath));
-        assertTrue(folderLocalStorageService.createFolder(USERNAME, destinationPath));
+        assertTrue(LocalStorageService.createFolder(USERNAME, sourcePath));
+        assertTrue(LocalStorageService.createFolder(USERNAME, destinationPath));
 
         Path source = Paths.get(storageLocation, USERNAME, sourcePath);
         Path dest = Paths.get(storageLocation, USERNAME, destinationPath);
@@ -287,7 +287,7 @@ public class FolderStorageServiceTest {
         assertTrue(Files.exists(dest));
 
         //when
-        assertTrue(folderLocalStorageService.move(USERNAME, source, dest));
+        assertTrue(LocalStorageService.move(USERNAME, source, dest));
 
         // then
         assertFalse(Files.exists(source));
@@ -305,8 +305,8 @@ public class FolderStorageServiceTest {
         Path sourcePath = Paths.get(origin);
         String destinationPath = "testFolder4";
 
-        assertTrue(folderLocalStorageService.createFolder(USERNAME, sourcePath.toString()));
-        assertTrue(folderLocalStorageService.createFolder(USERNAME, destinationPath));
+        assertTrue(LocalStorageService.createFolder(USERNAME, sourcePath.toString()));
+        assertTrue(LocalStorageService.createFolder(USERNAME, destinationPath));
 
         Path source = Paths.get(storageLocation, USERNAME, sourcePath.toString());
         Path dest = Paths.get(storageLocation, USERNAME, destinationPath);
@@ -316,11 +316,11 @@ public class FolderStorageServiceTest {
         Path testFolder5 = Paths.get(sourcePath.toString(), "testFolder5");
         Path testFolder6 = Paths.get(testFolder5.toString(), "testFolder6");
 
-        assertTrue(folderLocalStorageService.createFolder(USERNAME, testFolder5.toString()));
-        assertTrue(folderLocalStorageService.createFolder(USERNAME, testFolder6.toString()));
+        assertTrue(LocalStorageService.createFolder(USERNAME, testFolder5.toString()));
+        assertTrue(LocalStorageService.createFolder(USERNAME, testFolder6.toString()));
 
         //when
-        assertTrue(folderLocalStorageService.move(USERNAME, source, dest));
+        assertTrue(LocalStorageService.move(USERNAME, source, dest));
 
         // then
         assertFalse(Files.exists(source));
